Mobile first with typescript initial setup
