--- .gitpod.yml
@@ -1,13 +1,0 @@
-tasks:
-  - init: |
-      sudo apt-get update
-      sudo apt-get install -y android-sdk
-      export ANDROID_HOME=/path/to/android/sdk
-      echo "export ANDROID_HOME=/path/to/android/sdk" >> ~/.bashrc
-      echo "export PATH=\$PATH:\$ANDROID_HOME/tools:\$ANDROID_HOME/platform-tools" >> ~/.bashrc
-      npm install
-    command: npm start
-
-ports:
-  - port: 19000-19003
-    onOpen: ignore
