+++ package.json
@@ -1,48 +1,22 @@
 {
-  "name": "peer2peer_bill_splitter",
-  "main": "expo-router/entry",
-  "license": "0BSD",
+  "name": "peer2peer-bill-splitter",
   "version": "1.0.0",
+  "main": "index.ts",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.1.0",
-    "@react-navigation/native": "^7.1.6",
-    "expo": "53.0.9",
-    "expo-font": "~13.3.1",
-    "expo-linking": "~7.1.5",
-    "expo-router": "~5.0.7",
-    "expo-splash-screen": "~0.30.8",
+    "expo": "~53.0.11",
     "expo-status-bar": "~2.2.3",
-    "expo-system-ui": "~5.0.7",
-    "expo-web-browser": "~14.1.6",
-    "lucide-react": "*",
-    "lucide-react-native": "*",
     "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-native": "0.79.2",
-    "react-native-paper": "4.9.2",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0",
-    "react-native-svg": "15.11.2",
-    "react-native-web": "~0.20.0",
-    "jest-expo": "~53.0.5"
+    "react-native": "0.79.3"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~19.0.10",
-    "jest": "^29.2.1",
-    "jest-expo": "~53.0.4",
-    "react-test-renderer": "19.0.0",
     "typescript": "~5.8.3"
   },
   "private": true
